Get rid of ActiveSupport::EventReporter::Encoders module
It's a needlessly deep namespace. With a shorter namespace there's
no need for the `.encoder` lookup method thing. Just referencing
a constant is fine.

Also move encoders in their own files so we can test for msgpack
presence at load time rather than on every `encode` call.
